pyproject.toml
[tool.poetry]
name = "ebscatalog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
django-jazzmin = "^2.5.0"
drf-yasg = "^1.21.3"
psycopg2-binary = "^2.9.3"
django-auditlog = "^2.1.1"
django-json-widget = "^1.1.1"
beautifulsoup4 = "^4.11.1"
fake-useragent = "^0.1.11"
django-debug-toolbar = "^3.6.0"
django-bulk-update-or-create = "^0.3.0"
urllib3 = "^1.26.12"
chardet = "^5.0.0"
Faker = "^14.2.1"
coverage = "^6.4.4"
django-extensions = "^3.2.1"
python-slugify = "^6.1.2"
django-rest-framework-nested = "^0.0.1"
djangorestframework-simplejwt = "^5.2.0"
slugify = "^0.0.1"
djangorestframework-filters = "^0.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
